refactor: changed responsability location for local imports
