+++ src/features/register-player/game-styles/helpers/index.ts
@@ -1,0 +1,2 @@
+export { defaultValues } from './defaultValues'
+export { validationCheckBoxLimitation } from './validationCheckBoxLimitation'
