+++ src/features/register-player/game-styles/hooks/index.ts
@@ -1,0 +1,3 @@
+export { useFetchData } from './useFetchData'
+export { useFormGameStyles } from './useFormGameStyles'
+export { useSubmitGameStyles } from './useSubmitGameStyles'
