+++ src/features/register-player/game-styles/hooks/useFormGameStyles.ts
@@ -3,10 +3,10 @@ import { useForm } from 'react-hook-form'
 
 import useLocalStorageGetItem from '@/features/register-player/address-registration/hooks/useLocalStorageGetItem'
 
-import { defaultValues } from '../helpers/defaultValues'
+import { defaultValues } from '../helpers'
 import { FormGameStylesType, GameStyle } from '../types/gameStyles'
 import { FormGameStylesSchema } from '../validation'
-import { useFetchData } from './useFetchData'
+import { useFetchData } from '.'
 
 export function useFormGameStyles() {
   const {
