+++ src/features/register-player/game-styles/hooks/useSubmitGameStyles.ts
@@ -5,7 +5,7 @@ import { RegisterRoutes } from '@/services/routers'
 
 import { FormGameStylesType } from '../types/gameStyles'
 
-export default function useSubmitGameStyles() {
+export function useSubmitGameStyles() {
   const router = useRouter()
   return (values: FormGameStylesType) => {
     uselocalStorageSetItem('game_styles', values)
