+++ src/features/register-player/player-registration/FormStepOne.tsx
@@ -23,8 +23,7 @@ import {
 
 import FormText from '../FormText.json'
 import { FormFirstStep } from '.'
-import { useFormStepOne } from './hooks/useFormStepOne'
-import useSubmitStepOne from './hooks/useSubmitStepOne'
+import { useFormStepOne, useSubmitStepOne } from './hooks'
 
 export function FormStepOne() {
   const { nameLoading, form, handleSelectDate } = useFormStepOne()
