+++ src/features/register-player/player-registration/hooks/index.ts
@@ -1,0 +1,5 @@
+export { useFormStepOne } from './useFormStepOne'
+export { useSubmitStepOne } from './useSubmitStepOne'
+export { useGetNameByClerk } from './useGetNameByClerk'
+export { useSetValueDateOfBirth } from './useSetValueDateOfBirth'
+export { useStateDateOfBirth } from './useStateDateOfBirth'
