+++ src/features/register-player/player-registration/hooks/useSubmitStepOne.tsx
@@ -5,7 +5,7 @@ import { RegisterRoutes } from '@/services/routers'
 
 import { FormStepOneType } from '../validation'
 
-export default function useSubmitStepOne() {
+export function useSubmitStepOne() {
   const router = useRouter()
   return (values: FormStepOneType) => {
     const notEmptyPronounValues = {
